littlefs: rearrange code to facilitate porting
Export selected functions from lfs.c
Move internal structs/functions to separate file
Move block device operations to separate file
Move configuration parameters to separate header file
Use error codes from errno.h

JIRA: RTOS-686
